Split DigestInteractor into two new test files, also replaced it to only utilize the interactor for a basic test
